Plasmons.jl

Installing

Algorithm

Within real-space RPA (Random Phase Approximation) the calculation of polarizability matrix $\chi(\omega)$ amounts to evaluation of the following equation

\[\begin{aligned} \chi_{a,b} &= \langle a |\hat\chi| b \rangle = 2\cdot \sum_{i,j} \langle i |\hat G| j \rangle \langle j | a \rangle \langle a | i \rangle \langle i | b \rangle \langle b | j \rangle\; , \\ G_{i,j} &= \langle i | \hat G | j \rangle = \frac{f_i - f_j}{E_i - E_j - (\hbar\omega + i\eta)} \;. \end{aligned}\]

Here $E_i$s are energy eigenvalues, $\langle a | i \rangle = \psi_i (a)$ are energy eigenstates, $f_i = f(E_i)$ is the occupation of the $i$'th state given by the Fermi-Dirac distribution, $\eta$ is Landau damping, and $\omega$ is the frequency.

First of all, for a given temperature $T$ and chemical potential $\mu$, we compute occupation numbers $f(E_i)$.

Plasmons.fermidirac — Function

fermidirac(E; mu, kT) -> f

Return Fermi-Dirac distribution $f$ at energy E, chemical potential mu, and temperature kT. Note that kT is assumed to be temperature multiplied by the Boltzmann constant, i.e. physical dimension of kT is the same as E (e.g. electron-volts).

Next, we compute the matrix $G$. This is done by either Plasmons._g or Plasmons._g_blocks functions.

Plasmons._g — Function

_g(ħω, E; mu, kT) -> G

Warning

This is an internal function!

Compute matrix G

\[ G_{ij} = \frac{f(E_i) - f(E_j)}{E_i - E_j - \hbar\omega}\]

where $f$ is Fermi-Dirac distribution at chemical potential mu ($\mu$) and temperature kT ($k_B T$). E is a vector of eigenenergies. ħω is a complex frequency including Landau damping (i.e. $\hbar\omega + i\eta$). All arguments are assumed to be in energy units.

Sometimes one can further exploit the structure of $G$. For $E \ll \mu$ or $E \gg \mu$ Fermi-Dirac distribution is just a constant and $G$ goes to 0 for all $\omega$. Plasmons._g_blocks uses this fact to construct a block-sparse version of $G$. The reason why such a block-sparse version is useful will become apparent later.

Plasmons._g_blocks — Function

_g_blocks(ħω, E; mu, kT) -> (Gᵣ, Gᵢ)

Warning

This is an internal function!

Calculate matrix $G$ given a vector of eigenenergies E, chemical potential mu, and temperature kT. See _g for the definition of $G$.

Compared to _g this function applies to tricks:

G is split into real and imaginary parts Gᵣ and Gᵢ.

We exploit the "block-sparse" structure of G (see Plasmons._ThreeBlockMatrix).

Let us now turn to the computation of polarizability matrix elements $\chi_{a, b}$.